Added BigInt based UInt256
